Renamed u63 to Int
In order to provide other types, there are a few options. Macros would
quickly become annoying as there would need to be one per type, and so
generics seem better. Supporting arbitrary bit widths is desirable, and
so being able to move above 128 bits and the range of a u128 is another
goal. Therefore making a universal integer called simply Int is a good
way forward, and the u63 specific implementation has been renamed to
this in preparation for the functional changes.

Notably, the tests have not been updated yet, as we will use them to
specifically test a 63-bit integer when the changes are complete.